src/AzureSmartCost.Api/appSmartCost/requirements.txt
azure-identity>=1.15.0
azure-mgmt-costmanagement>=4.0.1
azure-mgmt-subscription>=3.1.1
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
